agents-workshop/03-building-custom-tools/03.5.on-behalf-of/pyproject.toml
[project]
name = "poc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "authlib>=1.3.2",
    "azure-monitor-opentelemetry-exporter>=1.0.0b36",
    "azure-search-documents>=11.6.0b4",
    "semantic-kernel[azure]>=1.31.0",
    "streamlit>=1.44.1",
]
